NODE-1088: Extract cli.py and abi.py
