--- integration-testing/test/test_python_client_bundled_contracts.py
@@ -1,5 +1,0 @@
-import casperlabs_client
-
-
-def test_check_bundled_contracts():
-    casperlabs_client.check_bundled_contracts()
